Version bumps for null safety support

+++ lib/main/utils/randomcolor/random_color/name_that_color.dart
@@ -1602,8 +1602,8 @@ MyColor getColorNameFromString(String color) {
   for (int i = 0; i < colorNames.length; i++) {
     if (color == colorNames[i].getCode) return colorNames[i];
 
-    ndf1 = pow(r - colorNames[i].getRed, 2) + pow(g - colorNames[i].getGreen, 2) + pow(b - colorNames[i].getBlue, 2);
-    ndf2 = pow(h - colorNames[i].getHue, 2) + pow(s - colorNames[i].getSaturation, 2) + pow(l - colorNames[i].getLightness, 2);
+    ndf1 = pow(r - colorNames[i].getRed, 2) + pow(g - colorNames[i].getGreen, 2) + pow(b - colorNames[i].getBlue, 2) as int;
+    ndf2 = pow(h - colorNames[i].getHue, 2) + pow(s - colorNames[i].getSaturation, 2) + pow(l - colorNames[i].getLightness, 2) as int;
     ndf = ndf1 + ndf2 * 2;
     if (df < 0 || df > ndf) {
       df = ndf;
